Resurrect portfolio.
This is largely pretty hacky as of right now, but it's a quick-and-dirty fix
until I have more time to rebuild in depth.

Uses custom post types and scans folders for HTML layouts. CSS is a mess since it's
copied more or less verbatim from old theme. Could use some love on mobile devices.