Set up post so we can apply individual styling to each.
We're adding the post slug to the array of post slugs so we can more easily
target the right posts, including between development and production environments.

We're also using lettering.js to wrap each word in a span, so we can target
the individual words. And finally, I've added some initial styling for "the
unbearable lightness of design" post, just to get a feel for it working.